backend/config/requirements.txt
fastapi
uvicorn
python-multipart
cryptography
python-dateutil
python-dotenv
PyJWT
cachetools
sqlalchemy